pyproject.toml
[project]
name = "ask-harry"
version = "0.1.0"
description = "A CLI tool for asking question related to your repo."
author = "ms-codehorizon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "ollama>=0.6.1",
    "pyspellchecker>=0.9.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

# Define the executable command name
[project.scripts]
ask-harry = "ask_harry.cli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/ask_harry"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
